Add a footer section
To improve the design of Vuebnb, let's add a footer to the bottom of
each page. We'll make it a reusable component. so let's begin by
creating that:

$ touch resources/assets/component/CustomFooter.vue

Let's add the footer to the App component, just below the RouterView
where the pages are output.

.container was specifically designed for the listing page, while
.home-container was designed for the home page. To fix this, and to make
thing less confusing, let's firstly rename the .container class to
.listing-container. You'll also need to update the ListingPage component
to ensure it's using this new class name.

Secondly, let's move .home-container to the main CSS file as well, since
we'll start to use it globally as well.

We have .home-container and .listing-container as two possible
containers for our custom-footer component. Let's dynamically select
class depending on the route, so the footer is always correctly aligned.