List summary image slider
On our home page, we need to make it so that a user can see more than
just three of the possible 10 listings for each country. To do this, we
will turn each listing summary group into an image slider.

Let's create a new component to house each listing summary group. We'll
then add arrowheads to the sides of this component, allowing the user to
easily step though its listings.

$ touch resources/assets/components/ListingSummaryGroup.vue

We'll abstract the markup and logic for displaying listing summaries
from HomePage into this new component. Each group will need to know the
name of the country and the included listings, so we'll add this data as
props.

Back in the HomePage, we will declare the ListingSummaryGroup with a
v-for, iterating over each country group.