Add the slider
We'll now add the slider functionality to ListingSummaryGroup. To do
this, we'll reuse the CarouselControl component. We'll want to display
one on either side of the group, so let's put them into the template,
remembering to declare the dir attribute. We'll also add some structural
markup and CSS for displaying the controls.