Use translate
In order to shift our listing summaries in response to the carousel
controls being clicked, we will use a CSS transform called translate.
This moves an affected element from its current position by an amount
specified in pixels.

The total width of each listing summary is 365px (350px fixed width plus
15px margin). This means if we move our group to the left by 365px, it
will give the effect of shifting the position of all images by one. You
can see here I've added the translate as inline styling to test if it
works. Note that we translate in a negative direction to get the group
to move to the left.

By binding inline style to the element with the listing-summary class,
we can control the translate from JavaScript. Let's do this via a
computed property so we can calculate the translate amount dynamically