Use carousel controls
We need the carousel controls to change the value of the translate. To
do so, let's add a data property, offset, to ListingSummaryGroup. This
will track how many images we've shifted along, that is, it will start
at zero, and go up to a maximun of seven (not 10 because we don't want
to shift so far along that all of the images are off-screen).

We also add a method change, which will serve as an event handling
function for the custom event that the carousel control components emit.
This method accepts one argument, val, which will either be -1 or 1,
depending on whether the left or right carousel control was triggered.

change will step the value of offset, which is then multiplied by the
width of each listing (365px) to calculate translate.

Lastly, we must use a v-on directive in the template to register a
listener for the change-image event of the CarouselControl components.