Finish touches
There are two more small features to add to these images slider to give
Vuebnb users the best possible experience. Firstly, let's add a CSS
transition to animate the translate change over a period of half a
second and give a nice sliding effect.

These sliders are not continuous; they have a minimum and maximum value.
Let's hide the appropiate arrow if that minimum of maximum is reached.
For example, when the slider load, the left arrow should be hidden
because the usr cannot decrement the offset further bellow zero.

We'll use style bindings to dynamically add a visibility: hidden CSS
rule and the computed properties.