Create store module js
We need to import Vuex in this file and, like Vue Router, install it
with Vue.use. This gives special properties to Vue that make it
compatible with Vuex, such as allowing components to access the store
via this.$store.

We import the store module in our main app file. and add it to our Vue
instance.