Create ListingSave component
The template of this component will include a Font Awesome heart icon.
It will also include a click handler which will be used to toggle tha
saved state. Since this component will always be a child of a listing of
listing summary, it will receive a listing ID  as a prop. This prop will
be used shortly to save the state in Vuex.

Note

The click handler has a stop modifier. This modifier prevents the click
event from bubbling up to ancestor elements, especially any anchor tags
which might trigger a page change!

We'll now add ListingSave to the ListingSummary component. Remember to
pass the the listing's ID as a prop. While we're at it, let's add a
position: relative to the .listing-summary class rules so that
ListingSave can be positioned absolutely against it.