Add saved state
The ListingSave component does not have any local data; we will instead
keep any saved listing in our Vuex store. To do this, we will create an
array in the store called saved. Each time the user toggles the saved
state of a listing its ID will be either added or removed from this
array.

To begin, let's add a state property to our Vuex store. This object will
hold any data we want to be globally available to the components of our
app. We will add the saved property to this object and assing it an
empty array.