Add a mutator method
We created the stub for a toggleSaved method in our ListingSave
component. This method should add or remove the listing's ID from the
saved state in the store. Components can access the store as
this.$store. More specifically, the saved array can be accessed at
this.$store.state.saved.

Remember that in the Flux architecture state is read-only. That means we
cannot directly modify saved from a component. Instead, we must create a
mutator method in the sotre which does the modification for us.

Let's create a mutations property in our store configuration, and add a
function property toggleSaved. Vuex mutator methods receive two
arguments: the store state and a payload. This payload can be anything
you want to pass from the component to the mutator. For the current
case, we will send the listing ID.

The logic for toggleSaved is to check if the listing ID is already in
the saved array and if so, remove it, or if not, add it.

We now need to commit this mutation from ListingSave. Commit is Flux
jargon that is synonymous with call or trigger. A commit looks like a
custom event with the first argument being the name of the mutator
method and the second being the payload.

The main point of using mutator methods in the store architecture is
that is changed consistently. But there is and additional benefit: we
can easily log these changes for debugging. If you check the Vuex tab in
Vue Devtools after clicking one of the save buttons, you'll see an entry
for that mutations.

Each entry in the log can tell you the state after the change was
commited, as well as the particulars of the mutation.

Note

If you double-click a logged mutation, Vue Devtools will revert the
state of the app to what it was directly after that change. This is
called time-travel debugging and can be useful for fine-grained
debugging.