Change the icon to reflect the state
Our ListingSave component's icon will appear differently, depending on
whether or not the listing is saved; it will be opaque if the listing is
saved, and transparent if it is not. Since the component doesn't store
its state locally, we need to retrieve state from the store to implement
this feature.

Vuex store state should generally be retrieve via a computed property.
This ensures that the component is not keeping its own copy, which would
violate the single source of truth principle, and that the component is
re-rendered when the state is mutated by this component or another.
Reativity works with Vuex state, too!

Let's create a computed property isListingSaved, which will return a
Boolean value reflecting whether or not this particular listing has been
saved.

We can now use this computed property to change the icon. Currently
we're using the Font Awesome icon fa-heart-o. This should represent the
unsaved state. When the listing is saved we should instead use the icon
fa-heart. We can implement this with a dynamic class binding.

Now the user can visually identify which listings have been saved and
which haven't. Thanks to reactive Vuex data, the icon will instantly be
updated when a change to the saved state is made from anywhere in the
app.