Add save feature to ListingPage
We also want the save feature to appear on the listing page. It will go
inside the HeaderImage component alongside the View Photos button so
that, like with the listing summaries, the button is overlaid on the
listing's main image.

Note that HeaderImage does not have the listing ID in its scope, so
we'll have to pass this down as a prop from ListingPage. id is not
currently a data property of ListingPage either, but, if we declared it,
it will simply work.

This is because the ID is already a property of the initial state/AJAX
data the component receives, therefore id will automatically be
populated by the Object.assign when the component is loaded by the
router.

Note

If you save a listing via the listing page, the return to the home page,
the equivalent listing summary will be saved. This is because our Vuex
state is global and will persist across page changes(though not page
refresh... yet).