Make ListingSave a button
The ListingSave feature appears too small in the listing page header and
will be easily overlooked by a user. Let's make it a proper button,
similar to the View Photos button in the bottom left of the header.

To do this, we'll modify ListingSave to allow parent components to send
a prop button. This Boolean prop will indicate if the component should
include a button element wrapped around the icon or not.

The text for this button will be a computed property message which will
change from Save to saved depending on the value of isListingSaved.

We will set the button prop to true within HeaderImage. Even though the
value is not dynamic, we use a v-bind to ensure the value is interpreted
as a JavaScript value, not a string.