Add listings & listing_summaries state properties
These will be arrays that store our listings and listing summaries
respectively. When the page first loads, or when the route changes and
the API is called, the loaded data will be put into these arrays rather
than being assigned directly to the page components. These page
components. The page components will instead retrieve this data from the
store.

We'll also add a mutator method, addData, for populating these arrays.
It will accept a payload object with two properties: route and
data. route is the name of the route, for example, listing, home, and so
on. data is the listing or listing summary data retrieved from the
document head or the API.