Store the state page on Vuex
The logic for retrieven page is in the mixin file route-mixin.js. This
mixin adds a beforeRouteEnter hook to a page component which applies the
page state to the component instance when it becomes available.

Now, that we're storing page state in Vuex we will utilize a different
approach. Firstly, we won't need a mixin anymore; we'll put this logic
into router.js now. Secondly, we'll use a different navigation guard,
beforeEach. This is not a component hook, but a hooj that can be applied
to the router itself, and it is triggered before every navigation.

You can see in the following code block how I've implemented this in
router.js. Note that before next() is called we commit the page state to
the store.

We delete the route mixin.

We've moved page state into Vuex we'll need to modify our page
components to retrieve it. Starting with ListingPage, the changes we
must make are:

- Remove local data properties
- Add a computed property listing. This will find the right listing data
from the store based on the route.
- Remove the mixin.
- Change template variables so they're properties of listing: an example
is {{ title }}, which will become {{ listing.title }}. Unfortunately,
all variables are now properties of listing which makes our template
slightly more verbose.

Changes to HomePage are much simpler; just remove the mixin and the
local state, and replace it with a computed property, listing_groups,
which will retrieve all the listing summaries from the store.

After making these changes, reload the app and you should see no obvious
change in behavior. However, inspecting the Vuex tab of Vue Devtools,
you will see that page data is now in the store.