Add getListing getter
You define a getter when several components need the same derived value,
to save repeating code. Let's create a getter which retrieves a specific
listing. We've already created this functionality in ListingPage, but
since we're going to need it in our router as well, we'll refactor it as
a getter.

One thing about getters is that they don't accept a payload argument
like mutations do. If you want to pass a value to a getter, you need to
return a function where the payload is an argument of that function.

Let's use the getListing getter in our ListingPage to replace the
previous logic.

Checkig if page state is in the store

Let's implement this logic in the beforeEach hook in router.js We'll add
an if block at the start that will instantly resolve the hook if the
data is already present. The if uses a ternary function with the
following logic.

- If the route name is listing, use the getListing getter to see if that
particular listing is available(this getter return undefined if it is
not)
- If the route name is not listing, check to see if the store has
listing summaries available. Listing summaries are always retrieved all
at once, so if there's at least one, you can asume they're all there.