Create Saved page
We will add the saved page to Vuebnb. Let's begin by creating the
component file.

$ touch resources/assets/components/SavedPage.vue

Next, we'll create a new route with this component at the path /saved.

Let's also add some server-side routes to the Laravel project. As
discussed above, the saved page uses exactly the same data as the home
page. This means that we can just call the same controller methods used
for the home page.

Let's define the SavePage component. We will import the
ListingSummaries component that we created. We'll also create a computed
property, listings, that will return the listing summaries from the
store, filtered by whether or not they're saved.

Next, we will add to the template tat of SavedPage. The main content
includes a check for the lenght of the array returned by the listings
computed property. If it is 0, no items have been yet. In this case , we
display a message to inform the user. If there are listing saved,
however, we'll iterate though them and display them with the
ListingSummary component.

We'll add to the style tag. The main thing to note here is that we're
utilizing the flex-wrap: wrap rule and justifying to the left. This
ensures that our listing summaries will organize themselves in rows
without gaps.

Let's also add the .saved-container CSS rules in our global CSS file.
This ensures that our custom footer has access to these rules as well.

The final task is to add some default saved listings to the store. I've
chosen 1 and 15 at random, but you can add any you want. We'll remove
these again later.