Add a toolber link
We add a link to saved page in the toolbar so that the saved page is
accessible from any other page. To do this, we'll add an inline ul where
links are enclosed within a child li.

To display correctly, we'll to add some extra CSS. Firstly, we'll modify
the #toolbar declaration so that the toolbar uses flex for display.
We'll also add some new rules below that for displaying the links.