Add the field saved to migration and model user class
Migration

We add an additional column for storing the saved listing IDs. Ideally,
we'd store these in an array, but since relationl database don't have an
array column type, we will instead store them as a serialized string,
for example, [1, 5, 10] within a text column.

Model

Let's allow the saved attribute to be mass assignable by adding it to
the $fillable array.

We'll also get our model to serialize and deserialize the saved text
when we read it or write it. To do this we can add a $casts attribute to
the model and cast saved as an array.

Now we can treat the saved attribute as an array, even though it's
stored as a string in the database.