Generate user account with database seeder
$ php artisan make:seeder UsersTableSeeder

Let's create at least one account with a name, mail, password, and an
array of saved listings. Note that I've used the make method of the hash
facade to hash the password rather than storing it as plain-text.
Laravel's default LoginController will automatically verify plain-text
passwords against the hash during the login process.

To run the seeder we need to call it from the main DatabaseSeeder class.

Let's now rerun our migrations and seeder to install the user table and
data with the following command:

$ php artisan migrate:refresh --seed

To confirm that our user table and data were created correctly, we'll
use Tinker to query the table.