Create LoginPage component
We will need a login page for our app, so let's create a new page
component.

$ touch resources/assets/components/LoginPage.vue

We'll begin by defining the template markup, which includes a form with
fields for email and password, and a submit button. The form uses the
HTTP POST method and is sent to the /login path. I've wrapped the form
elements in a div with the .form-controller class to help with styling.

We'll add a login-container class to our global CSS file so the footer
for this page align correctly. We'll also add a CSS rule to ensure text
inputs display correctly on iPhone. The login page will be the only
place we'll have a text input, but let's add it as a global rule in case
you decide to add other forms later.

Finnaly, let's add this new page component to our router. We'll first
import the component then add it to our routes array in the router
configuration.

Note that the login page does not require any data from the server like
other pages of Vuebnb do. This means that we can skip the data-fetching
step by modifying the logic of the first if statement in the navigation
guard. It should now resolve straightaway if the name of the route is
login.