Create a matching server-side route
We will also need a route for the login form that post to the same
/login path.

All we have to do to activate the routes is add the following line to
the bottom of our web route file.

To see the routes, we can use Artisan to show a list of the routes in
our app:

$ php artisan route:list

We'll see all routes that we've manually created, plus a few that we
didn't, for example, login, logout, and register. These are the routes
used by Laravel's authentication system that we just activated.

Looking at the GET/HEAD /login route, you'll see that it points to the
LoginController controller.

This class uses an AuthenticatesUsers trait that defines the
showLoginForm method that the /login route handler refers to. Let's
overwrite that method so it simply returns our app view. Since this
instance of the view doesn't need any data to be inlined in the head(the
login form has no state), we will pass an empty array to the data
template variable.

With that done, we can now see our complete login page by navigating the
browser to /login.