Add CSRF token and post login redirection
We add a CSRF token to the head of the page and assign it to the window
object.

We can now retrieve this from within our Vue.js app and manually add it
to the login form. Let's modify LoginPage to include a hidden input
field in the form. We'll add some state to this component now, in which
the token is included as a data property and bound to the hidden field.

When a user logs in, Laravel will redirect them to a page defined by the
$redirectTo attribute in the login controller. Let's change this from
/home to /.

Let's update the RedirectIfAuthenticated middleware class so that if a
logged-in user attempts to view the login page, they're redirect to /.