Add authentication links to the toolbar
Let's add login and logout links in the toolbar so Vuebnb users can
easily access these features.

The login link is simply a RouteLink pointing to the login route.

The logout link is a bit more interesting: we capture the click event
from this link and trigger the submission of a hidden form. This form
sends a POST request to the /logout server route, which logs the user
out and redirects them back to the home page. Note that we must include
the CSRF token as a hidden input form this to work.