Add auth meta property
We'll begin by adding the authentication state to the meta information
we pass through in the head of each page. We'll utilize the Auth facade
check method, which returns true if the user is authenticated, and
assign this to a new auth property.

We'll also add an auth property to our Vuex store. We'll mutate it from
the addDate method which, as you'll recall from the previous chapter, is
where we retrieve data from the document head or API. Since the API does
not include meta data, we'll conditionally mutate the auth property to
avoid accessing a potentially undefined object property.

With that done, Vuex is now tracking the authentication state of the
user. Be sure to test this out by logging in and out noticing the value
of auth in the Vuex tab of Vue Devtools.