Respond to authenticated state
We import our router module at the top of the file to access its
features.

We'll also make it so that either the login or the logout link is shown
in the toolbar, never both. This can be achieved by using v-if and
v-else directives in the toolbar that depend on the $store.state.auth
value.

It would also make sense to hide the saved page link unless the user is
logged in.