Retrieve saved items from the database
We add a new saved property to the metadata we put in the document head.
This will be an empty array if the user is logged out, or the array of
saved listing IDs associated with that user, if they're logged in.

We put the logic for retrieving the saved items in the beforeEach router
navigation guard. The reason we put it here and not in the addData
mutations is that we don't want to directly assign the data to the store
state, but instead call the toggleSaved mutation for each of the
listings. You can't commit a mutation from another mutation, so this
must be done outside the store.

Let's also remove the placeholder listing IDs we added to saved property.