Create an API route
Add a route for the frontend to POST listing IDS to. We'll need to add
the auth middleware so that only authenticated users can access this
route.

Let's create a UserController:

$ php artisan make:controller UserController

We add the toggle_saved handling method. Since this is an HTTP POST
route, this method will have access to the form data. We'll make it so
that the frontend AJAX call to this route includes an id field, which
will be the listing ID we want to toggle. To access this field, we can
use the Input facade, that is, Input::get('id');.

Since we're using the auth middleware on this route, we can retrieve the
user model associated with the request by using the Auth::user() method.
We can then either add or remove the ID from the user's saved listing,
just as we do in the toggledSaved method in our Vuex store.

Once the ID is toggled, we can then use the model's save method to
persist the update to the database.