Use AJAX to use toggle_save api controller method
Let's use AJAX to make the request to /api/user/toggle_saved when a
listing is saved. We'll put this logic into a Vuex action so that the
toggleSaved mutation is not committed until the AJAX call resolves.
We'll import the Axios HTTP library into the store to facilitate this.

Let's move the authentication check from the mutation to the action, as
it makes sense to do this check before the AJAX call is initiated.

We call the toggleSaved action, from our ListingSave component.