Attach tokens
OAuth requires an access token to be sent to the frontend app when the
user logs in. Passport includes a middleware that can handle this for
you. Add the CreateFreshApiToken middleware to the web middleware group
and the laravel_token cookie will be attached to outgoing response.

For outgoing requests, we need to add some headers to our AJAX calls. We
can make it so Axios automatically attaches these by default.
'X-Requested-With': 'XMLHttpRequest' ensures that Laravel knows the
request was from AJAX, while 'X-CSRF-TOKEN': window.csrf_token attaches
the CSRF token.

With that done, our API request should now be properly authenticated. To
test this, let's use Tinker to see which items we have saved for our
first seeded user:

$ php artisan tinker

    DB::table('users')->select('saved')->first();